Switch S3Client.get_object_meta to a bulk method
We may be dealing with thousands of files for some jobs, so better to
share a client and parallelize, which is the next step after switching
the method to take multiple paths.

Also makes a paths container class that handles path validation, so we
can create it immediately upon reception of paths from the client and
pass it around without having to revalidate. Will also have at least one
more method that takes a set of paths, which means the validation tests
don't need to be run again for that method.