Make S3Client batch methods concurrent
The results are actually a little disappointing, only 50% faster when
`head`ing 1k files. That being said, the code is done, it doesn't add
much complexity, 50% faster is 50% faster, and when running against a
beefier system than a single minio instance on my laptop that can handle
more simultaneous connections the performance improvement may be more
substantial.

```
writing files
done
testing with concurrency of 1000
testing with concurrency of 100
testing with concurrency of 10
testing with concurrency of 1
Concur	Time	Improvement
1	1.56064	1.000%
10	1.07180	1.456%
100	1.10735	1.409%
1000	1.04405	1.495%
```