Switch the nersc token provider to a client provider
The token provider can provide a token that lasts for at most 10 minutes
and could last for as little as one minute based on the default refresh
configuration. This means that if the SFAPI client needs to poll for
more than that amount of time it'll start getting 403s.

Instead, provide the client and allow it to manage token refreshes while
polling.

The client provider will
* want to cache the client so it doesn't recreate it every call, since
that causes a call to the token endpoint
* need to deal with the fact that the credentials expire periodically,
at most after 30 days

Also made the chmod code only add 100 instead of 111.