Download files to NERSC on job submission
Job submission code kicks off a coroutine that runs a task at NERSC to
do the download. This allows the submission code to return to the user
as soon as sanity checks are complete and state is saved.

Had to switch the download target from the DTNs to the login node since
for some reason `$SCRATCH` and `module` aren't available.

Also had to switch the docker compose to host network node so that the
service could access the ssh tunnel to Minio.