Fix missing shell quote
When adding the prefix & suffix to container number, missed adding a
shell quote. Previously wasn't needed since it was just a number.

Also remove note about supporting globbing, since shell quoting disables
globbing. May want to rethink this if there's a call for globbing as we
currently have 2 layers of protection - the restricted characters in
models.py and the shell quoting in wdl.py.