Mongo DAO refactor, part 1
The driver for this is that for every job update we need to send a
message to mongo, and soon to kafka as well. This is the first step of
doing that in only one place in the codebase instead of, currently, 10.
It abstracts a job update into a class based representation so that
we'll only need 1 mongo method that accepts the abstraction to perform
the update instead of the multitude we have. The job flows will create a
job update representation with the code in the new module and send that
to the single mongo method called in one place. To add kafka we just add
the kafka call right after the mongo call.