Fix Image de/serialization in mongo
`mode="json"` is the wrong way to go since it serializes datetimes to
strings, which we don't want. Instead make a custom serializer for
HttpUrls, which was causing mongo to choke.

Similarly, use model_validate instead of model_construct so that the
strings are turned into HttpUrls correctly.

Also a minor doc fix